requirements.txt
numpy
unidecode
librosa
soundfile
herpetologist
dataclasses
python_speech_features
unidecode
tqdm
tornado
scipy>=1.5.4
malaya-boilerplate>=0.0.24rc1
transformers
torchlibrosa==0.0.7